SchoolAi/backend/requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
matplotlib
seaborn
textblob
tensorflow
Pillow
nltk

